Refactor LogBusHandler - move from ro2018.py to bus.py
