+++ osgar/ro2018.py
@@ -15,7 +15,7 @@ from osgar.drivers import all_drivers
 from osgar.robot import Robot
 
 from osgar.drivers.gps import INVALID_COORDINATES
-from osgar.drivers.bus import BusHandler
+from osgar.drivers.bus import BusHandler, LogBusHandler, LogBusHandlerInputsOnly
 
 
 def geo_length(pos1, pos2):
@@ -197,48 +197,6 @@ class RoboOrienteering2018:
         while self.time - start_time < timedelta(seconds=3):
             self.set_speed(0, 0)
             self.update()
-
-
-# move to drivers/bus??
-class LogBusHandler:
-    def __init__(self, log, inputs, outputs):
-        self.reader = log.read_gen(list(inputs.keys()) + list(outputs.keys()))
-        self.inputs = inputs
-        self.outputs = outputs
-        self.buffer_queue = Queue()
-
-    def listen(self):
-        if self.buffer_queue.empty():
-            dt, stream_id, data = next(self.reader)
-        else:
-            dt, stream_id, data = self.buffer_queue.get()
-        channel = self.inputs[stream_id]
-        return dt, channel, literal_eval(data.decode('ascii'))
-
-    def publish(self, channel, data):
-        assert channel in self.outputs.values(), (channel, self.outputs.values())
-        dt, stream_id, raw_data = next(self.reader)
-        while stream_id not in self.outputs:
-            assert stream_id in self.inputs, stream_id
-            self.buffer_queue.put((dt, stream_id, raw_data))
-            dt, stream_id, raw_data = next(self.reader)
-        assert channel == self.outputs[stream_id], (channel, self.outputs[stream_id])  # wrong channel
-        ref_data = literal_eval(raw_data.decode('ascii'))
-        assert data == ref_data, (data, ref_data)
-
-
-class LogBusHandlerInputsOnly:
-    def __init__(self, log, inputs):
-        self.reader = log.read_gen(inputs.keys())
-        self.inputs = inputs
-
-    def listen(self):
-        dt, stream_id, data = next(self.reader)
-        channel = self.inputs[stream_id]
-        return dt, channel, literal_eval(data.decode('ascii'))
-
-    def publish(self, channel, data):
-        pass
 
 
 if __name__ == "__main__":
